\name{D310}
\docType{data}
\alias{D310}
\title{Roquemore (1976) Hybrid design D310}
\description{
    A
	This is an .rda file containing the design in a matrix.
}
\usage{data(D310)}
\format{Three columns of independent variables}
\source{source}
\references{
    Myers, R. H. and Montgomery D. C. \emph{Response Surface Methodology} 2nd Ed., John Wiley and Sons NY, 2002. p.386
}
\keyword{datasets}
