\name{D416A}
\docType{data}
\alias{D416A}
\title{Roquemore (1976) Hybrid design 416A}
\description{
    This is an .rda file containing the design in a matrix.
}
\usage{data(D416A)}
\format{Four columns of independent variables}
\source{source}
\references{
    Myers, R. H. and Montgomery D. C. \emph{Response Surface Methodology} 2nd Ed., John Wiley and Sons NY, 2002. p.386
}
\keyword{datasets}
