\name{D416C}
\docType{data}
\alias{D416C}
\title{Roquemore (1976) Hybrid design D416C}
\description{
    This is an .rda file containing the design in a matrix.
}
\usage{data(D416C)}
\format{Three columns of independent variables}
\source{source}
\references{
    Myers, R. H. and Montgomery D. C. \emph{Response Surface Methodology} 2nd Ed., John Wiley and Sons NY, 2002. p.386
}
\keyword{datasets}
