\name{f}
\alias{f}
\title{ Calculate column means of design}
\description{This function calculates means of design. }
\usage{  
f(x)
}
\arguments{
  \item{x}{ This is a design matrix}
}
\value{mean
\item{mean }{ This is the mean of the design x }
}

\note{ This function is called by the function Vdgraph.}
\author{ John S. Lawson \email{lawson@byu.edu}}
\keyword{datagen}
