\name{Hex2}
\docType{data}
\alias{Hex2}
\title{Hexagonal design for two factors}
\description{
    This is an .rda file containing the design in a matrix.
}
\usage{data(Hex2)}
\format{Two columns of independent variables}
\source{source}
\references{
    Myers, R. H. and Montgomery D. C. \emph{Response Surface Methodology} 2nd Ed., John Wiley and Sons NY, 2002. p.386
}
\keyword{datasets}
