\name{mx}
\alias{mx}
\title{ Calculate column maximums of design}
\description{This function calculates maximums of design. }
\usage{  
mx(x)
}
\arguments{
  \item{x}{ This is a design matrix}
}
\value{mean
\item{max }{ This is the maximum of the design x }
}

\note{ This function is called by the function FDSPlot.}
\author{ John S. Lawson \email{lawson@byu.edu}}
\keyword{datagen}
