\name{SCDH3}
\docType{data}
\alias{SCDH3}
\title{Hartley's Small Composite Design for three factors}
\description{
    This is an .rda file containing the design in a matrix.
}
\usage{data(SCDH3)}
\format{Three columns of independent variables}
\source{source}
\references{
    Myers, R. H. and Montgomery D. C. \emph{Response Surface Methodology} 2nd Ed., John Wiley and Sons NY, 2002. p.386
}
\keyword{datasets}
