\name{SCDH6}
\docType{data}
\alias{SCDH6}
\title{Hartley's Small Composite Design for six factors}
\description{
    This is an .rda file containing the design in a matrix.
}
\usage{data(SCDH6)}
\format{Six columns of independent variables}
\source{source}
\references{
    Myers, R. H. and Montgomery D. C. \emph{Response Surface Methodology} 2nd Ed., John Wiley and Sons NY, 2002. p.386
}
\keyword{datasets}
