\name{SCDDL5}
\docType{data}
\alias{SCDDL5}
\title{Draper and Lin's Small Composite Design for five factors}
\description{
    An .rda file containing the design in a matrix.
}
\usage{data(SCDDL5)}
\format{Five columns of independent variables}
\source{source}
\references{
    Myers, R. H. and Montgomery D. C. \emph{Response Surface Methodology} 2nd Ed., John Wiley and Sons NY, 2002. p.386
}
\keyword{datasets}
